backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# OpenAI API
openai>=1.12.0

# Vector database
qdrant-client>=1.8.0

# PostgreSQL (using psycopg3 - compatible with Python 3.14)
psycopg[binary]>=3.2.0

# Database migrations
alembic>=1.13.1
sqlalchemy>=2.0.25

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# Additional utilities
python-multipart==0.0.6
email-validator==2.1.0
